Move proxy to src
